Haciendo refactor de funciones comunes a utilizar en los componentes de grupos
