--- frontend/www/js/omegaup/group/edit.test.ts
@@ -1,17 +1,0 @@
-import { generatePassword, generateHumanReadablePassword } from './edit';
-
-describe('group_edit', () => {
-  describe('generatePassword', () => {
-    const simplePasswordLength = 8;
-
-    it('Should be human-readable (greater than 8 characters)', () => {
-      const password = generateHumanReadablePassword();
-      expect(password.length).toBeGreaterThan(simplePasswordLength);
-    });
-
-    it("Shouldn't be human-readable (equal to 8 characters)", () => {
-      const password = generatePassword();
-      expect(password.length).toBe(simplePasswordLength);
-    });
-  });
-});
