+++ frontend/www/js/omegaup/groups_utils.ts
@@ -1,0 +1,105 @@
+import T from './lang';
+
+export function cleanRecords(
+  records: (null | number | string)[][],
+): (undefined | string)[][] {
+  return records.map((row) =>
+    row.map((cell) => {
+      if (cell === null) {
+        return undefined;
+      }
+      if (typeof cell !== 'string') {
+        return String(cell);
+      }
+      return cell;
+    }),
+  );
+}
+
+export function generatePassword(): string {
+  const validChars = 'acdefhjkmnpqruvwxyACDEFHJKLMNPQRUVWXY346';
+  const len = 8;
+  // Browser supports window.crypto
+  if (typeof window.crypto == 'object') {
+    const arr = new Uint8Array(2 * len);
+    window.crypto.getRandomValues(arr);
+    return Array.from(
+      arr.filter((value) => value <= 255 - (255 % validChars.length)),
+      (value) => validChars[value % validChars.length],
+    )
+      .join('')
+      .substr(0, len);
+  }
+
+  // Browser does not support window.crypto
+  let password = '';
+  for (let i = 0; i < len; i++) {
+    password += validChars.charAt(
+      Math.floor(Math.random() * validChars.length),
+    );
+  }
+  return password;
+}
+
+export function generateHumanReadablePassword() {
+  const words = {
+    es: [
+      'Loro',
+      'Perro',
+      'Pollo',
+      'Lagarto',
+      'Gato',
+      'Toro',
+      'Vaca',
+      'Sapo',
+      'Oso',
+      'Zorro',
+    ],
+    en: [
+      'Parrot',
+      'Dog',
+      'Chicken',
+      'Lizard',
+      'Cat',
+      'Bull',
+      'Cow',
+      'Frog',
+      'Bear',
+      'Fox',
+    ],
+    pt: [
+      'Papagaio',
+      'Cachorro',
+      'Frango',
+      'Lagarto',
+      'Gato',
+      'Touro',
+      'Vaca',
+      'Sapo',
+      'Urso',
+      'Raposa',
+    ],
+  };
+  const wordsNumber = 12;
+  const totalNumbers = 6;
+
+  let langWords: string[] = [];
+  switch (T.locale) {
+    case 'es':
+      langWords = words.es;
+      break;
+    case 'pt':
+      langWords = words.pt;
+      break;
+    default:
+      langWords = words.en;
+  }
+  let password = '';
+  for (let i = 0; i < wordsNumber; i++) {
+    password += langWords[Math.floor(Math.random() * langWords.length)];
+  }
+  for (let i = 0; i < totalNumbers; i++) {
+    password += Math.floor(Math.random() * 10); // random numbers
+  }
+  return password;
+}
